refactor: keep only Search component
