+++ src/lib/Scenes/BottomTabs/BottomTabsNavigator.tsx
@@ -29,7 +29,7 @@ export const BottomTabsNavigator = () => {
         <FadeBetween
           views={[
             <TabContent key="home" route={{ params: { tabName: "home", rootModuleName: "Home" } }} />,
-            <TabContent key="search" route={{ params: { tabName: "search", rootModuleName: "Search2" } }} />,
+            <TabContent key="search" route={{ params: { tabName: "search", rootModuleName: "Search" } }} />,
             <TabContent key="inbox" route={{ params: { tabName: "inbox", rootModuleName: "Inbox" } }} />,
             <TabContent key="sell" route={{ params: { tabName: "sell", rootModuleName: "Sales" } }} />,
             <TabContent key="profile" route={{ params: { tabName: "profile", rootModuleName: "MyProfile" } }} />,
@@ -47,7 +47,7 @@ export const BottomTabsNavigator = () => {
       <Tab.Screen
         name="search"
         component={TabContent}
-        initialParams={{ tabName: "search", rootModuleName: "Search2" }}
+        initialParams={{ tabName: "search", rootModuleName: "Search" }}
       />
       <Tab.Screen name="inbox" component={TabContent} initialParams={{ tabName: "inbox", rootModuleName: "Inbox" }} />
       <Tab.Screen name="sell" component={TabContent} initialParams={{ tabName: "sell", rootModuleName: "Sales" }} />
