+++ src/lib/Scenes/Search/RecentSearches.tests.tsx
@@ -2,7 +2,7 @@ import { __globalStoreTestUtils__, GlobalStore } from "lib/store/GlobalStore"
 import { extractText } from "lib/tests/extractText"
 import { renderWithWrappers } from "lib/tests/renderWithWrappers"
 import React from "react"
-import { AutosuggestSearchResult } from "./AutosuggestSearchResult"
+import { AutosuggestSearchResult } from "./components/AutosuggestSearchResult"
 import { RecentSearches } from "./RecentSearches"
 import { SearchContext } from "./SearchContext"
 import { RecentSearch } from "./SearchModel"
