+++ src/lib/Scenes/Search/SearchResults.tsx
@@ -6,7 +6,7 @@ import { Box, Flex, Spacer, Spinner, Text, useSpace } from "palette"
 import React, { useEffect, useRef, useState } from "react"
 import { InfiniteHitsProvided, StateResultsProvided } from "react-instantsearch-core"
 import { FlatList } from "react-native"
-import { AlgoliaSearchPlaceholder } from "./components/AlgoliaSearchPlaceholder"
+import { AlgoliaSearchPlaceholder } from "./components/placeholders/AlgoliaSearchPlaceholder"
 import { SearchResult } from "./components/SearchResult"
 import { isAlgoliaApiKeyExpiredError } from "./helpers"
 import { AlgoliaSearchResult } from "./types"
