+++ src/lib/Scenes/Search/components/SearchPills.tests.tsx
@@ -1,8 +1,8 @@
 import { fireEvent } from "@testing-library/react-native"
 import { renderWithWrappersTL } from "lib/tests/renderWithWrappers"
 import React from "react"
-import { SearchPills, SearchPillsProps } from "./components/SearchPills"
-import { PillType } from "./types"
+import { PillType } from "../types"
+import { SearchPills, SearchPillsProps } from "./SearchPills"
 
 const pills: PillType[] = [
   {
