+++ src/lib/Scenes/Search/components/SearchResult.tsx
