+++ src/lib/Scenes/Search/components/SearchResultImage.tsx
