+++ src/lib/Scenes/Search/components/placeholders/AutosuggestResultsPlaceholder.tsx
@@ -2,7 +2,7 @@ import { PlaceholderBox, RandomWidthPlaceholderText } from "lib/utils/placeholde
 import { times } from "lodash"
 import { Box, Flex } from "palette"
 import React from "react"
-import { IMAGE_SIZE } from "../Search2/components/SearchResultImage"
+import { IMAGE_SIZE } from "../SearchResultImage"
 
 const TEXT_SIZE = 12
