+++ src/lib/Scenes/Search/helpers.ts
